+++ Back_End/src/main/java/com/ssafy/youniverse/dto/res/GenreResDto.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class GenreResDto {
-    private Integer genreId;
+    private Long genreId;
     private String genreName;
 }
