+++ Back_End/src/main/java/com/ssafy/youniverse/dto/res/RecommendOttResDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class RecommendOttResDto {
-    private Integer ottId;
+    private Long ottId;
     private String ottName;
     private String ottImage;
     private String ottUrl;
