+++ Back_End/src/main/java/com/ssafy/youniverse/entity/HateMovie.java
@@ -1,11 +1,9 @@
 package com.ssafy.youniverse.entity;
 
-import lombok.Data;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import javax.persistence.*;
 
 @Getter
 @Setter
@@ -15,7 +13,7 @@ public class HateMovie {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer hateMovieId;
+    private Long hateMovieId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
