+++ Back_End/src/main/java/com/ssafy/youniverse/entity/YoutubeKeyword.java
@@ -1,12 +1,11 @@
 package com.ssafy.youniverse.entity;
 
 import com.ssafy.youniverse.util.Auditable;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Getter
 @Setter
@@ -15,7 +14,7 @@ import javax.persistence.*;
 public class YoutubeKeyword extends Auditable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer youtubeKeywordId;
+    private Long youtubeKeywordId;
 
     @Column(nullable = false, length = 30)
     private String youtubeKeywordName;
