+++ Back_End/src/main/java/com/ssafy/youniverse/repository/ActorMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.ActorMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActorMovieRepository extends JpaRepository<ActorMovie, Integer> {
+public interface ActorMovieRepository extends JpaRepository<ActorMovie, Long> {
 }
