+++ Back_End/src/main/java/com/ssafy/youniverse/repository/GenreMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.GenreMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenreMovieRepository extends JpaRepository<GenreMovie, Integer> {
+public interface GenreMovieRepository extends JpaRepository<GenreMovie, Long> {
 }
