+++ Back_End/src/main/java/com/ssafy/youniverse/repository/HateMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.HateMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HateMovieRepository extends JpaRepository<HateMovie, Integer> {
+public interface HateMovieRepository extends JpaRepository<HateMovie, Long> {
 }
