+++ Back_End/src/main/java/com/ssafy/youniverse/repository/HeartMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.HeartMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HeartMovieRepository extends JpaRepository<HeartMovie, Integer> {
+public interface HeartMovieRepository extends JpaRepository<HeartMovie, Long> {
 }
