+++ Back_End/src/main/java/com/ssafy/youniverse/repository/MovieRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer>, MovieRepositoryCustom {
+public interface MovieRepository extends JpaRepository<Movie, Long>, MovieRepositoryCustom {
 
 }
