+++ Back_End/src/main/java/com/ssafy/youniverse/repository/RecommendMovieRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.youniverse.repository;
 import com.ssafy.youniverse.entity.RecommendMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecommendMovieRepository extends JpaRepository<RecommendMovie, Integer> {
+public interface RecommendMovieRepository extends JpaRepository<RecommendMovie, Long> {
 }
